Move QueryFactory under Query.
